+++ pyccel/ast/builtins.py
@@ -1207,13 +1207,15 @@ class PythonType(PyccelAstNode):
         then be easily printed in each language.
         """
         prec = self.precision
-        dtype = str(self.dtype)
-        if prec in (None, -1):
+        dtype = self.dtype.name
+        if prec in (None, 0, -1):
             return LiteralString(f"<class '{dtype}'>")
 
+        class_type = self._obj.class_type
+
         precision = prec * (16 if self.dtype is NativeComplex() else 8)
-        if self._obj.rank > 0:
-            return LiteralString(f"<class 'numpy.ndarray' ({dtype}{precision})>")
+        if class_type != self.dtype:
+            return LiteralString(f"<class '{class_type}' ({dtype}{precision})>")
         else:
             return LiteralString(f"<class 'numpy.{dtype}{precision}'>")