+++ pyccel/ast/datatypes.py
@@ -147,7 +147,7 @@ class NativeBool(DataType, metaclass=Singleton):
     Class representing a boolean datatype.
     """
     __slots__ = ()
-    _name = 'Bool'
+    _name = 'bool'
 
     @lru_cache
     def __add__(self, other):
@@ -163,7 +163,7 @@ class NativeInteger(DataType, metaclass=Singleton):
     Class representing an integer datatype.
     """
     __slots__ = ()
-    _name = 'Int'
+    _name = 'int'
 
     @lru_cache
     def __add__(self, other):
@@ -182,7 +182,7 @@ class NativeFloat(DataType, metaclass=Singleton):
     Class representing a float datatype.
     """
     __slots__ = ()
-    _name = 'Float'
+    _name = 'float'
 
     @lru_cache
     def __add__(self, other):
@@ -201,7 +201,7 @@ class NativeComplex(DataType, metaclass=Singleton):
     Class representing a complex datatype.
     """
     __slots__ = ()
-    _name = 'Complex'
+    _name = 'complex'
 
     @lru_cache
     def __add__(self, other):
@@ -220,7 +220,7 @@ class NativeString(DataType, metaclass=Singleton):
     Class representing a string datatype.
     """
     __slots__ = ()
-    _name = 'String'
+    _name = 'str'
 
     @lru_cache
     def __add__(self, other):
@@ -238,7 +238,7 @@ class NativeVoid(DataType, metaclass=Singleton):
     pointers from Fortran.
     """
     __slots__ = ()
-    _name = 'Void'
+    _name = 'void'
 
 class NativeTuple(DataType):
     """
@@ -255,7 +255,7 @@ class NativeTuple(DataType):
         Any keyword arguments required by the class.
     """
     __slots__ = ()
-    _name = 'Tuple'
+    _name = 'tuple'
 
     @lru_cache
     def __add__(self, other):
@@ -322,7 +322,7 @@ class NativeHomogeneousList(DataType, metaclass = Singleton):
     is a container type and should be used as the class_type.
     """
     __slots__ = ()
-    _name = 'List'
+    _name = 'list'
 
     @lru_cache
     def __add__(self, other):